refactored upload to use API class
